Commit 10: Prim's Minimum Spanning Tree Method
Implemented Prim's Spanning Tree Algorithm and added tests in main. Basically Prim's algorithm is about subset of the edges that is going to be minimum in cost, but connect all the vertices and there is no circle. So ultimately we want to form a tree which is going to be a connected component and this tree is going to have minimum cost. There might be multiple minimal cost trees.